Add MBC3 cartridge type
